identity/requirements.txt
aioboto3
python-dotenv